backend/requirements.txt
# Web框架
fastapi==0.115.0
uvicorn[standard]==0.30.0
pydantic==2.9.0
pydantic-settings==2.5.0

# 数据库
sqlalchemy[asyncio]==2.0.35
asyncpg==0.29.0
alembic==1.13.0

# 认证
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Redis
redis[hiredis]==5.1.0

# MQTT客户端
paho-mqtt==1.6.1

# 工具
python-multipart==0.0.12
openpyxl==3.1.5
httpx==0.27.0

# 任务调度
apscheduler==3.10.4
